refactor: move irepo and repo to the correct namespace
