+++ Modules/Tokens/src/Tokens.Application/Infrastructure/IUnitOfWork.cs
@@ -1,4 +1,6 @@
-﻿namespace Backbone.Modules.Tokens.Application.Infrastructure;
+﻿using Backbone.Modules.Tokens.Application.Infrastructure.Persistence.Repository;
+
+namespace Backbone.Modules.Tokens.Application.Infrastructure;
 
 public interface IUnitOfWork : IDisposable
 {
