+++ Modules/Tokens/src/Tokens.Infrastructure/Persistence/Database/Repository/IServiceCollectionExtensions.cs
@@ -1,8 +1,9 @@
 ﻿using Backbone.Modules.Tokens.Application.Infrastructure;
+using Backbone.Modules.Tokens.Application.Infrastructure.Persistence.Repository;
 using Enmeshed.BuildingBlocks.Infrastructure.Persistence.BlobStorage;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace Backbone.Modules.Tokens.Infrastructure.Persistence.Repository;
+namespace Backbone.Modules.Tokens.Infrastructure.Persistence.Database.Repository;
 
 public static class IServiceCollectionExtensions
 {
