+++ Modules/Tokens/test/Tokens.Application.Tests/FakeUnitOfWork.cs
@@ -1,4 +1,5 @@
 ﻿using Backbone.Modules.Tokens.Application.Infrastructure;
+using Backbone.Modules.Tokens.Application.Infrastructure.Persistence.Repository;
 
 namespace Backbone.Modules.Tokens.Application.Tests;
